requirement.txt
pycryptodome
requests